Views & more models logic (not finished)

+++ core/src/main/java/com/krolewskie_potyczki/model/Tower.java
@@ -1,7 +1,7 @@
 package com.krolewskie_potyczki.model;
 
 public class Tower extends Entity {
-    Tower(boolean isPlayersEntity, double x, double y) {
+    Tower(boolean isPlayersEntity, float x, float y) {
         super(isPlayersEntity, x, y);
     }
 }
